+++ nodebackend/src/clients/vaultClient.ts
@@ -1,6 +1,6 @@
-import * as envSwitcher from './envSwitcher';
-import { vaultRoleId, vaultSecretId } from './configs';
-import logger from './logger';
+import * as envSwitcher from '../envSwitcher';
+import { vaultRoleId, vaultSecretId } from '../configs';
+import logger from '../logger';
 import vault from 'node-vault';
 
 const vaultClient = vault({