+++ nodeserver/routes/loginRoute.js
@@ -1,9 +1,9 @@
 const express = require('express');
 const router = express.Router();
 const { loginValidation } = require('../inputValidation');
-const { connectToRedis } = require('../../nodebackend/build/redisClient');
+const { connectToRedis } = require('../../nodebackend/build/clients/redisClient');
 const crypto = require('crypto');
-const vaultClient = require('../../nodebackend/build/vaultClient'); // Import your Vault client
+const vaultClient = require('../../nodebackend/build/clients/vaultClient'); // Import your Vault client
 const logger = require('../../nodebackend/build/logger').default;
 
 router.post('/', async (req, res) => {