+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] 2023-11-26
+## [v15.2.1] 2023-12-02
 ### Changed
 - Migration of influxdb-client to TS
 - Migration of traditonalcheck.js to TS